Lösungsklasse mit Kostenberechnung hinzugefügt. Abbruchkriterium für Savings-Algorithmus eingefügt. Logische Fehler behoben, sodass Savings-Algorithmus nun korrekt funktioniert.

+++ simultane-planung/src/model/Journey.java
@@ -6,7 +6,7 @@ public interface Journey {
 
 	public void setId(String id);
 
-	public int getDistance();
+	public double getDistance();
 	
 	public String getFromStopId();
